Finish components + make viewer open to students
