+++ portal/templates/portal/partials/hero_card.html
@@ -8,11 +8,11 @@
             <h3 class="card__title">{{ title }}</h3>
             <p>{{ description }}</p>
             <div class="button-group">
-                <a href="{% url button1_link %}" class="button button--primary--general-play">
-                    {{ button1_text }}<div class="glyphicon glyphicon-menu-right"></div>
+                <a href="{% url button1.url pdf_name=button1.url_args %}" class="button button--primary--general-play">
+                    {{ button1.text }}<div class="glyphicon glyphicon-menu-right"></div>
                 </a>
-                <a href="{% url button2_link %}" class="button button--primary--general-play">
-                    {{ button2_text }}<div class="glyphicon glyphicon-menu-right"></div>
+                <a href="{% url button2.url id=button2.url_args %}" class="button button--primary--general-play">
+                    {{ button2.text }}<div class="glyphicon glyphicon-menu-right"></div>
                 </a>
             </div>
         </div>
